+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - `Moar.Duration.to_short_string/1` converts duration to short strings like `"3s"`.
 - Add `Moar.Duration.humanize/1` which can turn `{120, :second}` into `{2, :minute}`,
   and `Moar.Duration.shift/1` which is like `Moar.Duration.convert/1` but returns a duration tuple.
+- Add `Moar.Duration.ago/1` which returns the duration between a given datetime and now.
+- Add `Moar.Duration.between/2` which returns the duration between two datetimes.
 
 ## 1.11.0